# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 4.0.0
  description: TIAMO BURGER — API del Sistema POS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /usuarios:
    get:
      operationId: getUsuarios
      tags: [usuarios]
      summary: Listar usuarios (solo admin)
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioPublico"
    post:
      operationId: crearUsuario
      tags: [usuarios]
      summary: Crear nuevo usuario (solo admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearUsuarioInput"
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPublico"
  /usuarios/{id}:
    put:
      operationId: actualizarUsuario
      tags: [usuarios]
      summary: Actualizar usuario (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarUsuarioInput"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioPublico"
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: eliminarUsuario
      tags: [usuarios]
      summary: Eliminar usuario (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Usuario eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
  /productos:
    get:
      operationId: getProductos
      tags: [productos]
      summary: Listar productos del catálogo
      parameters:
        - name: categoria
          in: query
          required: false
          schema:
            type: string
        - name: soloDisponibles
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Producto"
    post:
      operationId: crearProducto
      tags: [productos]
      summary: Crear producto (solo admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearProductoInput"
      responses:
        "201":
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
  /productos/{id}:
    put:
      operationId: actualizarProducto
      tags: [productos]
      summary: Actualizar producto (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarProductoInput"
      responses:
        "200":
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Producto"
        "404":
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: eliminarProducto
      tags: [productos]
      summary: Eliminar producto (solo admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Producto eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RolUsuario:
      type: string
      enum: [admin, mesero, cocinero, caja]
...
    UsuarioPublico:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        usuario:
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        activo:
          type: boolean
        creadoEn:
          type: string
          format: date-time
        fotoUrl:
          type: string
          nullable: true
      required: [id, nombre, usuario, rol, activo, creadoEn]
...
    CrearUsuarioInput:
      type: object
      properties:
        nombre:
          type: string
        usuario:
          type: string
        contrasena:
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
      required: [nombre, usuario, contrasena, rol]
    ActualizarUsuarioInput:
      type: object
      properties:
        nombre:
          type: string
        contrasena:
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        activo:
          type: boolean
...
    Producto:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: string
        precio:
          type: integer
        emoji:
          type: string
        imagenUrl:
          type: string
        categoria:
          type: string
        disponible:
          type: boolean
        destacado:
          type: boolean
        posicion:
          type: integer
        creadoEn:
          type: string
          format: date-time
      required: [id, nombre, precio, emoji, categoria, disponible, destacado, posicion, creadoEn]
    CrearProductoInput:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        precio:
          type: integer
        emoji:
          type: string
        imagenUrl:
          type: string
        categoria:
          type: string
        disponible:
          type: boolean
        destacado:
          type: boolean
        posicion:
          type: integer
      required: [nombre, precio, categoria]
    ActualizarProductoInput:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        precio:
          type: integer
        emoji:
          type: string
        imagenUrl:
          type: string
        categoria:
          type: string
        disponible:
          type: boolean
        destacado:
          type: boolean
        posicion:
          type: integer
...